requirements.txt
torch>=2.0,<2.2
torchvision>=0.15,<0.17
mmpose==1.3.2
mmcv-lite>=2.0.0
mmdet>=3.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-image==0.21.0
pandas==2.0.3
tqdm==4.66.1
pyyaml==6.0.1
matplotlib==3.7.2
pytest==7.4.3
pytest-cov==4.1.0
